Create the code of this management
I create the code of the book and inventory Inquiry System in
order to have a platform for customers to search the books they
want to buy, and how many books still have in the book store.
Furthermore, the staff can add and delete books on the web store
that can facilitate real-time update of books' inventory data. I
create a Book Class to add the basic information of books, and the BookManage Class to store all books' information. Using five methods to create five functions of the Inquiry system.